requirements.txt
python-dotenv>=0.19.0
openai>=0.27.0
youtube-transcript-api>=0.4.0
aiohttp>=3.8.0
asyncio==3.4.3
streamlit>=1.32.0 